Rename : 스플래쉬 SplashTemplate 폴더로 이동 및 SymbolLogoCard 코드 이동
